requirements.txt
streamlit
requests
PIL
os
base64numpy
